Export ModalProps type
I realised that the ModalProps type was the same as what was being required in the Modal component so now it is being exported. Additionally I added the ability to set custom titles in the modal so that different product types can have different text displayed in the body and the title reflects that.