Fix checkout so that it works as intended:tm:
Previously the store would allow payments, process them correctly but wouldn't actually be able to reward  items based on the purchase. I changed the entire store checkout to use invoices rather than charges so that this works. The downside to this is that we are unable to process both 'one_time' products and 'recurring' products at the same time.